CHORE: Rename files for clarity
