+++ cleanlab/datalab/internal/issue_manager/__init__.py
@@ -4,5 +4,5 @@ from .label import LabelIssueManager
 from .outlier import OutlierIssueManager
 from .noniid import NonIIDIssueManager
 from .imbalance import ClassImbalanceIssueManager
-from .underperf_group import UnderperformingGroupIssueManager
+from .underperforming_group import UnderperformingGroupIssueManager
 from .data_valuation import DataValuationIssueManager
